requirements.txt
# LangGraph & LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Neo4j
neo4j>=5.0.0

# PostgreSQL
psycopg2-binary>=2.9.0

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Streamlit UI
streamlit>=1.40.0
requests>=2.32.0

# Data / Visualisation
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.20.0

# Utilities
python-dotenv>=1.0.0
